rfp-automation/requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
anthropic>=0.7.0
langchain>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.20
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pydantic>=2.0.0
typing-extensions>=4.5.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=0.8.11